Add controllers, services and repositoryes for users and items

+++ src/main/java/ru/practicum/shareit/user/service/UserServiceImp.java
@@ -4,6 +4,8 @@ import lombok.RequiredArgsConstructor;
 import lombok.SneakyThrows;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
+import ru.practicum.shareit.exception.ObjectNotFoundException;
+import ru.practicum.shareit.exception.UserNameDuplicateException;
 import ru.practicum.shareit.user.dto.UserDto;
 import ru.practicum.shareit.user.mapper.UserMapper;
 import ru.practicum.shareit.user.model.User;
@@ -25,6 +27,10 @@ public class UserServiceImp implements UserService {
     @Override
     public UserDto createUser(UserDto userDto) {
         log.info("Попытка добавить нового пользователя");
+        if (userRepository.emailContains(userDto.getEmail())) {
+            log.info("Пользователь с email:{} существует", userDto.getEmail());
+            throw new UserNameDuplicateException("Пользователь с email:" + userDto.getEmail() + " уже зарегестрирован");
+        }
         final User user = UserMapper.toUser(userDto);
         customValidator.isUserValid(user);
         return UserMapper.toUserDto(userRepository.createUser(user));
@@ -33,7 +39,21 @@ public class UserServiceImp implements UserService {
     @SneakyThrows
     @Override
     public UserDto updateUser(UserDto userDto, Long userId) {
-        log.info("Попытка обновить информацию о пользователе");
+        log.info("Попытка обновить информацию о пользователе id:{}", userId);
+        if (!userRepository.contains(userId)) {
+            log.info("Попытка обновить несуществующего пользователя");
+            throw new ObjectNotFoundException("Такой пользователь не зарегестрирован");
+        }
+        if (userRepository.emailContains(userDto.getEmail())) {
+            log.info("Пользователь с email:{} существует", userDto.getEmail());
+            throw new UserNameDuplicateException("Пользователь с email:" + userDto.getEmail() + " уже зарегестрирован");
+        }
+        if (userDto.getName() == null) {
+            return UserMapper.toUserDto(userRepository.updateUserWithoutName(UserMapper.toUser(userDto), userId));
+        }
+        if (userDto.getEmail() == null) {
+            return UserMapper.toUserDto(userRepository.updateUserWithoutEmail(UserMapper.toUser(userDto), userId));
+        }
         return UserMapper.toUserDto(userRepository.updateUser(UserMapper.toUser(userDto), userId));
     }
 
@@ -41,6 +61,10 @@ public class UserServiceImp implements UserService {
     @Override
     public String deleteUser(Long userId) {
         log.info("Попытка удалить пользователя");
+        if (!userRepository.contains(userId)) {
+            log.info("Попытка удалить несуществующего пользователя c id:{}", userId);
+            throw new ObjectNotFoundException("Пользователь c id:" + userId + " не зарегестрирован");
+        }
         return userRepository.deleteUser(userId);
     }
 
@@ -48,6 +72,10 @@ public class UserServiceImp implements UserService {
     @Override
     public UserDto findUserById(Long userId) {
         log.info("Попытка получить информацию о пользователе");
+        if (!userRepository.contains(userId)) {
+            log.info("Попытка получить информацию о несуществующем пользователе c id:{}", userId);
+            throw new ObjectNotFoundException("Пользователь c id:" + userId + " не зарегестрирован");
+        }
         return UserMapper.toUserDto(userRepository.findUserById(userId));
     }
